feat(mobile): ask for the native language once, on the way in
New accounts landed in the library with 'I know: English / Learning:
English' and nativeLanguage null on the server, so the reader's core
feature translated English into English. The question is now a step
between registering and the library.

One question, not two. The QA report asked for both halves, but
TARGET_LANGUAGES is NATIVE_LANGUAGES.filter(code === 'en') — a question
with one possible answer is not a question, it is the dead 'Learning'
chip the same report filed separately as P3-4.

No skip. Skipping restores the exact state the screen exists to end: a
guessed value indistinguishable from a chosen one. It is not a dead end
either — English is one tap at the top of Popular. The difference is that
it becomes an answer instead of a silence.

Two entry points, one rule. Registration routes here directly; a root
gate catches accounts made before the screen existed and reinstalls where
storage is empty but the profile is old. Both call shouldAskForLanguage,
because two copies of this rule would drift and both failure modes are
invisible in review.

The list moves out of LanguagePickerModal into LanguageList so the modal
and the screen share one body. Copying it would have been the fourth time
this codebase grew a second implementation of something it already had.

The strings were already written — en.json has had an onboarding block
with 'Choose your language' since before any of this. Adds two keys that
name the actual question instead of the generic one.
