+++ apps/mobile/src/components/LanguagePickerModal.tsx
@@ -1,15 +1,9 @@
-import { useState, useMemo, useEffect } from 'react'
-import { View, Text, StyleSheet, TouchableOpacity, Modal, Pressable, FlatList, TextInput } from 'react-native'
+import { useRef, useEffect } from 'react'
+import { View, Text, StyleSheet, TouchableOpacity, Modal, Pressable } from 'react-native'
 import { Ionicons } from '@expo/vector-icons'
 import { useTheme } from '../context/ThemeContext'
 import { fonts } from '../theme/typography'
-import {
-  LANGUAGES,
-  POPULAR_LANGUAGES,
-  OTHER_LANGUAGES,
-  getFlagEmoji,
-  type LanguageEntry,
-} from '../data/languages'
+import { LanguageList, type LanguageListHandle } from './LanguageList'
 
 interface Props {
   visible: boolean
@@ -18,36 +12,13 @@ interface Props {
   onChange: (code: string) => void
 }
 
-type Row =
-  | { kind: 'section'; title: string }
-  | { kind: 'item'; lang: LanguageEntry }
-
 export function LanguagePickerModal({ visible, onClose, value, onChange }: Props) {
   const { colors } = useTheme()
-  const [query, setQuery] = useState('')
+  const listRef = useRef<LanguageListHandle>(null)
 
   useEffect(() => {
-    if (!visible) setQuery('')
+    if (!visible) listRef.current?.reset()
   }, [visible])
-
-  const rows: Row[] = useMemo(() => {
-    const q = query.trim().toLowerCase()
-    if (!q) {
-      return [
-        { kind: 'section', title: 'Popular' },
-        ...POPULAR_LANGUAGES.map((lang) => ({ kind: 'item' as const, lang })),
-        { kind: 'section', title: 'All languages' },
-        ...OTHER_LANGUAGES.map((lang) => ({ kind: 'item' as const, lang })),
-      ]
-    }
-    const filtered = LANGUAGES.filter(
-      (l) =>
-        l.englishName.toLowerCase().includes(q) ||
-        l.nativeName.toLowerCase().includes(q) ||
-        l.code.toLowerCase().startsWith(q),
-    )
-    return filtered.map((lang) => ({ kind: 'item' as const, lang }))
-  }, [query])
 
   const handleSelect = (code: string) => {
     onChange(code)
@@ -79,87 +50,7 @@ export function LanguagePickerModal({ visible, onClose, value, onChange }: Props
             </TouchableOpacity>
           </View>
 
-          <View style={[styles.searchRow, { borderColor: colors.border }]}>
-            <Ionicons name="search" size={18} color={colors.textSecondary} style={{ marginRight: 8 }} />
-            <TextInput
-              value={query}
-              onChangeText={setQuery}
-              placeholder="Search language..."
-              placeholderTextColor={colors.textSecondary}
-              style={[styles.searchInput, { color: colors.text, fontFamily: fonts.sans }]}
-              autoCorrect={false}
-              autoCapitalize="none"
-              autoFocus
-              accessibilityLabel="Search language"
-              returnKeyType="search"
-            />
-          </View>
-
-          {rows.length === 0 ? (
-            <Text style={[styles.empty, { color: colors.textSecondary }]}>
-              No results for "{query.trim()}"
-            </Text>
-          ) : (
-            <FlatList
-              data={rows}
-              keyExtractor={(item, i) =>
-                item.kind === 'section' ? `section-${item.title}` : `item-${item.lang.code}-${i}`
-              }
-              keyboardShouldPersistTaps="handled"
-              style={styles.list}
-              renderItem={({ item }) => {
-                if (item.kind === 'section') {
-                  return (
-                    <Text
-                      style={[styles.sectionLabel, { color: colors.textSecondary }]}
-                      accessibilityRole="header"
-                    >
-                      {item.title}
-                    </Text>
-                  )
-                }
-                const lang = item.lang
-                const selected = lang.code === value
-                return (
-                  <TouchableOpacity
-                    style={[
-                      styles.row,
-                      { borderBottomColor: colors.border },
-                      selected && { backgroundColor: colors.primaryLight },
-                    ]}
-                    onPress={() => handleSelect(lang.code)}
-                    activeOpacity={0.7}
-                    accessibilityRole="button"
-                    accessibilityLabel={
-                      lang.englishName === lang.nativeName
-                        ? lang.nativeName
-                        : `${lang.nativeName}, ${lang.englishName}`
-                    }
-                    accessibilityState={{ selected }}
-                  >
-                    <Text style={styles.flag}>{getFlagEmoji(lang.code)}</Text>
-                    <View style={{ flex: 1 }}>
-                      <Text
-                        style={[
-                          styles.rowNative,
-                          { color: colors.text, fontFamily: fonts.sansMedium },
-                          selected && { color: colors.primary },
-                        ]}
-                      >
-                        {lang.nativeName}
-                      </Text>
-                      {lang.englishName !== lang.nativeName && (
-                        <Text style={[styles.rowEnglish, { color: colors.textSecondary, fontFamily: fonts.sans }]}>
-                          {lang.englishName}
-                        </Text>
-                      )}
-                    </View>
-                    {selected && <Ionicons name="checkmark" size={20} color={colors.primary} />}
-                  </TouchableOpacity>
-                )
-              }}
-            />
-          )}
+          <LanguageList ref={listRef} value={value} onSelect={handleSelect} autoFocusSearch />
         </Pressable>
       </Pressable>
     </Modal>
@@ -195,36 +86,4 @@ const styles = StyleSheet.create({
   },
   title: { fontSize: 17, fontFamily: fonts.sansBold },
   closeBtn: { padding: 4 },
-  searchRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderRadius: 8,
-    paddingHorizontal: 12,
-    paddingVertical: 8,
-    marginBottom: 8,
-  },
-  searchInput: { flex: 1, fontSize: 15, padding: 0 },
-  list: { flex: 1 },
-  sectionLabel: {
-    fontSize: 11,
-    textTransform: 'uppercase',
-    letterSpacing: 0.8,
-    fontFamily: fonts.sansMedium,
-    paddingVertical: 8,
-    paddingHorizontal: 4,
-  },
-  row: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingVertical: 12,
-    paddingHorizontal: 8,
-    borderBottomWidth: 1,
-    gap: 12,
-    borderRadius: 4,
-  },
-  flag: { fontSize: 22 },
-  rowNative: { fontSize: 15 },
-  rowEnglish: { fontSize: 12, marginTop: 2 },
-  empty: { textAlign: 'center', padding: 24, fontSize: 14 },
 })
